pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "satisfaculty"
version = "0.1.0"
description = "A course scheduling optimization tool using integer linear programming"
authors = [{name = "Zachary Sunberg", email = "zachary.sunberg@colorado.edu"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = "~=3.10"
classifiers = [
    "License :: OSI Approved :: MIT License",
]
keywords = ["scheduling", "optimization", "integer-programming", "course-scheduling"]
dependencies = [
    "pandas~=2.3",
    "numpy~=2.4",
    "PuLP~=3.3",
    "matplotlib~=3.10",
    "python-dotenv~=1.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest",
]

[project.urls]
Repository = "https://github.com/zsunberg/satisfaculty"

[tool.setuptools.packages.find]
include = ["satisfaculty*"]
